# pqarrow: change back to row-by-row conversion but write directly