@asubiotto 2023-11-28
Also added another commit to avoid calling `GetOneForMarshal` on int32/64 types since that allocates